standardize my code onto the same structure as guilherme's
