+++ x/emissions/module/rewards_v2_internal_test.go
@@ -1,4 +1,4 @@
-package equations
+package module
 
 import (
 	"math"
